Move ConsolidateBuffer to timely-util
